requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.8.0
Django==5.2
django-environ==0.12.0
django-extensions==4.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
Faker==37.1.0
iniconfig==2.1.0
kombu==5.5.2
packaging==24.2
pillow==11.2.1
pluggy==1.5.0
prompt_toolkit==3.0.51
PyJWT==2.9.0
pytest==8.3.5
python-dateutil==2.9.0.post0
redis==5.2.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
gunicorn==21.2.0
whitenoise==6.6.0
django-cors-headers==4.3.1
